Test cleanup.
